requirements.txt
google_api_python_client==2.136.0
google_auth_oauthlib==1.2.0
python-dotenv==1.0.1
youtube_transcript_api==0.6.2
